+++ src/main/java/com/nhnacademy/illuwa/order/client/MemberOrderClient.java
@@ -1,13 +1,13 @@
 package com.nhnacademy.illuwa.order.client;
 
 import com.nhnacademy.illuwa.common.dto.PageResponse;
-import com.nhnacademy.illuwa.order.dto.OrderCreateResponse;
-import com.nhnacademy.illuwa.order.dto.OrderListResponse;
-import com.nhnacademy.illuwa.order.dto.OrderResponse;
-import com.nhnacademy.illuwa.order.dto.member.MemberOrderCartRequest;
-import com.nhnacademy.illuwa.order.dto.member.MemberOrderDirectRequest;
-import com.nhnacademy.illuwa.order.dto.member.MemberOrderInitDirectResponse;
-import com.nhnacademy.illuwa.order.dto.member.MemberOrderInitFromCartResponse;
+import com.nhnacademy.illuwa.order.dto.query.info.OrderCreateResponse;
+import com.nhnacademy.illuwa.order.dto.query.list.OrderListResponse;
+import com.nhnacademy.illuwa.order.dto.query.detail.OrderResponse;
+import com.nhnacademy.illuwa.order.dto.command.create.MemberOrderCartRequest;
+import com.nhnacademy.illuwa.order.dto.command.create.MemberOrderDirectRequest;
+import com.nhnacademy.illuwa.order.dto.query.info.MemberOrderInitDirectResponse;
+import com.nhnacademy.illuwa.order.dto.query.info.MemberOrderInitFromCartResponse;
 import jakarta.validation.Valid;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.*;
