+++ src/main/java/com/nhnacademy/illuwa/order/dto/command/create/MemberOrderDirectRequest.java
@@ -1,49 +1,37 @@
-package com.nhnacademy.illuwa.order.dto.guest;
+package com.nhnacademy.illuwa.order.dto.command.create;
 
-import com.nhnacademy.illuwa.order.dto.orderRequest.OrderItemDto;
+import com.nhnacademy.illuwa.order.dto.query.detail.OrderItemResponse;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.PositiveOrZero;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class GuestOrderDirectRequest {
+public class MemberOrderDirectRequest {
     @NotNull
     private LocalDate deliveryDate;
-
     @NotBlank
     private String recipientName;
-
     @NotBlank
     private String recipientContact;
-
     @NotBlank
     private String postCode;
-
     @NotBlank
     private String readAddress;
-
     @NotBlank
     private String detailAddress;
-
     @NotNull
     @Valid
-    private OrderItemDto item;
-
-    @NotBlank
-    private String name;
-
-    @NotBlank
-    @Email
-    private String email;
-
-    @NotBlank
-    private String contact;
+    private OrderItemResponse item;
+    private Long memberCouponId;
+    @NotNull
+    @PositiveOrZero
+    private BigDecimal usedPoint;
 }
