+++ src/main/java/com/nhnacademy/illuwa/order/dto/query/detail/OrderItemResponse.java
@@ -1,22 +1,27 @@
-package com.nhnacademy.illuwa.order.dto;
+package com.nhnacademy.illuwa.order.dto.query.detail;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
 
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class OrderItemResponseDto {
-    private Long orderItemId;
-    private String title;
+public class OrderItemResponse {
+
+    @JsonProperty("id")
     private Long bookId;
-    private int quantity;
+
+    private Integer quantity;
+
     private BigDecimal price;
+
     private Long packagingId;
+
+    private Long couponId;
+
     private BigDecimal totalPrice;
-    private PackagingResponseDto packaging;
 }
