+++ src/main/java/com/nhnacademy/illuwa/order/dto/query/detail/OrderResponse.java
@@ -1,22 +1,26 @@
-package com.nhnacademy.illuwa.order.dto;
+package com.nhnacademy.illuwa.order.dto.query.detail;
 
-import com.nhnacademy.illuwa.order.dto.orderRequest.OrderItemDto;
 import com.nhnacademy.illuwa.order.dto.types.OrderStatus;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class OrderListResponse {
-
+public class OrderResponse {
     private Long orderId;
     private String orderNumber;
-    private LocalDateTime orderDate;
+    private Long memberId; // 회원 넘버
+    private LocalDateTime orderDate; //
+    private LocalDate deliveryDate;
+    private BigDecimal shippingFee;
     private BigDecimal totalPrice;
     private OrderStatus orderStatus;
+    private List<OrderItemResponseDto> items;
 }
