+++ src/main/java/com/nhnacademy/illuwa/order/dto/query/info/GuestOrderInitDirectResponse.java
@@ -1,0 +1,15 @@
+package com.nhnacademy.illuwa.order.dto.query.info;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class GuestOrderInitDirectResponse {
+    private BookPriceResponse item;
+    private List<PackagingResponse> packaging;
+}
