+++ src/main/java/com/nhnacademy/illuwa/order/dto/query/info/PackagingResponse.java
@@ -1,11 +1,11 @@
-package com.nhnacademy.illuwa.order.dto;
+package com.nhnacademy.illuwa.order.dto.query.info;
 
 import lombok.Data;
 
 import java.math.BigDecimal;
 
 @Data
-public class PackagingResponseDto {
+public class PackagingResponse {
     private Long packagingId;
     private String packagingName;
     private BigDecimal packagingPrice;
