+++ src/main/java/com/nhnacademy/illuwa/order/dto/query/list/OrderListResponse.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.illuwa.order.dto;
+package com.nhnacademy.illuwa.order.dto.query.list;
 
 import com.nhnacademy.illuwa.order.dto.types.OrderStatus;
 import lombok.AllArgsConstructor;
@@ -6,21 +6,16 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class OrderResponse {
+public class OrderListResponse {
+
     private Long orderId;
     private String orderNumber;
-    private Long memberId; // 회원 넘버
-    private LocalDateTime orderDate; //
-    private LocalDate deliveryDate;
-    private BigDecimal shippingFee;
+    private LocalDateTime orderDate;
     private BigDecimal totalPrice;
     private OrderStatus orderStatus;
-    private List<OrderItemResponseDto> items;
 }
