+++ src/main/java/com/nhnacademy/illuwa/order/service/AdminOrderService.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.illuwa.order.service;
 
 import com.nhnacademy.illuwa.common.dto.PageResponse;
-import com.nhnacademy.illuwa.order.dto.OrderListResponse;
-import com.nhnacademy.illuwa.order.dto.OrderResponse;
+import com.nhnacademy.illuwa.order.dto.query.list.OrderListResponse;
+import com.nhnacademy.illuwa.order.dto.query.detail.OrderResponse;
 import com.nhnacademy.illuwa.order.dto.types.OrderStatus;
 
 public interface AdminOrderService {
