+++ src/main/java/com/nhnacademy/illuwa/order/service/MemberOrderService.java
@@ -1,13 +1,13 @@
 package com.nhnacademy.illuwa.order.service;
 
 import com.nhnacademy.illuwa.common.dto.PageResponse;
-import com.nhnacademy.illuwa.order.dto.OrderCreateResponse;
-import com.nhnacademy.illuwa.order.dto.OrderListResponse;
-import com.nhnacademy.illuwa.order.dto.OrderResponse;
-import com.nhnacademy.illuwa.order.dto.member.MemberOrderCartRequest;
-import com.nhnacademy.illuwa.order.dto.member.MemberOrderDirectRequest;
-import com.nhnacademy.illuwa.order.dto.member.MemberOrderInitDirectResponse;
-import com.nhnacademy.illuwa.order.dto.member.MemberOrderInitFromCartResponse;
+import com.nhnacademy.illuwa.order.dto.query.info.OrderCreateResponse;
+import com.nhnacademy.illuwa.order.dto.query.list.OrderListResponse;
+import com.nhnacademy.illuwa.order.dto.query.detail.OrderResponse;
+import com.nhnacademy.illuwa.order.dto.command.create.MemberOrderCartRequest;
+import com.nhnacademy.illuwa.order.dto.command.create.MemberOrderDirectRequest;
+import com.nhnacademy.illuwa.order.dto.query.info.MemberOrderInitDirectResponse;
+import com.nhnacademy.illuwa.order.dto.query.info.MemberOrderInitFromCartResponse;
 
 public interface MemberOrderService {
