+++ src/main/java/com/nhnacademy/illuwa/order/service/PackagingService.java
@@ -1,16 +1,16 @@
 package com.nhnacademy.illuwa.order.service;
 
 
-import com.nhnacademy.illuwa.order.dto.PackagingRequestDto;
-import com.nhnacademy.illuwa.order.dto.PackagingResponseDto;
+import com.nhnacademy.illuwa.order.dto.command.create.PackagingRequest;
+import com.nhnacademy.illuwa.order.dto.query.info.PackagingResponse;
 
 import java.util.List;
 
 public interface PackagingService {
 
     // 활성화된 포장 옵션 조회
-    List<PackagingResponseDto> getAllPackagingOptions();
+    List<PackagingResponse> getAllPackagingOptions();
 
     // 새로운 포장 옵션 생성
-    void createPackaging(PackagingRequestDto dto);
+    void createPackaging(PackagingRequest dto);
 }
