+++ src/main/java/com/nhnacademy/illuwa/order/service/impl/AdminOrderServiceImpl.java
@@ -2,9 +2,9 @@ package com.nhnacademy.illuwa.order.service.impl;
 
 import com.nhnacademy.illuwa.common.dto.PageResponse;
 import com.nhnacademy.illuwa.order.client.AdminOrderClient;
-import com.nhnacademy.illuwa.order.dto.OrderListResponse;
-import com.nhnacademy.illuwa.order.dto.OrderResponse;
-import com.nhnacademy.illuwa.order.dto.admin.OrderUpdateStatusDto;
+import com.nhnacademy.illuwa.order.dto.query.list.OrderListResponse;
+import com.nhnacademy.illuwa.order.dto.query.detail.OrderResponse;
+import com.nhnacademy.illuwa.order.dto.command.status.OrderUpdateStatusDto;
 import com.nhnacademy.illuwa.order.dto.types.OrderStatus;
 import com.nhnacademy.illuwa.order.service.AdminOrderService;
 import lombok.RequiredArgsConstructor;
