+++ src/main/java/com/nhnacademy/illuwa/order/service/impl/GuestOrderServiceImpl.java
@@ -1,10 +1,10 @@
 package com.nhnacademy.illuwa.order.service.impl;
 
 import com.nhnacademy.illuwa.order.client.GuestOrderClient;
-import com.nhnacademy.illuwa.order.dto.OrderCreateResponse;
-import com.nhnacademy.illuwa.order.dto.OrderResponse;
-import com.nhnacademy.illuwa.order.dto.guest.GuestOrderDirectRequest;
-import com.nhnacademy.illuwa.order.dto.guest.GuestOrderInitDirectResponse;
+import com.nhnacademy.illuwa.order.dto.query.info.OrderCreateResponse;
+import com.nhnacademy.illuwa.order.dto.query.detail.OrderResponse;
+import com.nhnacademy.illuwa.order.dto.command.create.GuestOrderDirectRequest;
+import com.nhnacademy.illuwa.order.dto.query.info.GuestOrderInitDirectResponse;
 import com.nhnacademy.illuwa.order.service.GuestOrderService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
