+++ src/main/java/com/nhnacademy/illuwa/order/service/impl/MemberOrderServiceImpl.java
@@ -2,13 +2,13 @@ package com.nhnacademy.illuwa.order.service.impl;
 
 import com.nhnacademy.illuwa.common.dto.PageResponse;
 import com.nhnacademy.illuwa.order.client.MemberOrderClient;
-import com.nhnacademy.illuwa.order.dto.OrderCreateResponse;
-import com.nhnacademy.illuwa.order.dto.OrderListResponse;
-import com.nhnacademy.illuwa.order.dto.OrderResponse;
-import com.nhnacademy.illuwa.order.dto.member.MemberOrderCartRequest;
-import com.nhnacademy.illuwa.order.dto.member.MemberOrderDirectRequest;
-import com.nhnacademy.illuwa.order.dto.member.MemberOrderInitDirectResponse;
-import com.nhnacademy.illuwa.order.dto.member.MemberOrderInitFromCartResponse;
+import com.nhnacademy.illuwa.order.dto.query.info.OrderCreateResponse;
+import com.nhnacademy.illuwa.order.dto.query.list.OrderListResponse;
+import com.nhnacademy.illuwa.order.dto.query.detail.OrderResponse;
+import com.nhnacademy.illuwa.order.dto.command.create.MemberOrderCartRequest;
+import com.nhnacademy.illuwa.order.dto.command.create.MemberOrderDirectRequest;
+import com.nhnacademy.illuwa.order.dto.query.info.MemberOrderInitDirectResponse;
+import com.nhnacademy.illuwa.order.dto.query.info.MemberOrderInitFromCartResponse;
 import com.nhnacademy.illuwa.order.service.MemberOrderService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
