+++ src/main/java/com/nhnacademy/illuwa/order/service/impl/PackagingServiceImpl.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.illuwa.order.service.impl;
 
 import com.nhnacademy.illuwa.order.client.PackagingClient;
-import com.nhnacademy.illuwa.order.dto.PackagingRequestDto;
-import com.nhnacademy.illuwa.order.dto.PackagingResponseDto;
+import com.nhnacademy.illuwa.order.dto.command.create.PackagingRequest;
+import com.nhnacademy.illuwa.order.dto.query.info.PackagingResponse;
 import com.nhnacademy.illuwa.order.service.PackagingService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,12 +16,12 @@ public class PackagingServiceImpl implements PackagingService {
     private final PackagingClient packagingClient;
 
     @Override
-    public List<PackagingResponseDto> getAllPackagingOptions() {
+    public List<PackagingResponse> getAllPackagingOptions() {
         return packagingClient.getAll();
     }
 
     @Override
-    public void createPackaging(PackagingRequestDto dto) {
+    public void createPackaging(PackagingRequest dto) {
         packagingClient.create(dto);
     }
 }
